update test and schema
