+++ src/documents/invoices/InvoiceTypeCodes.ts
@@ -50,7 +50,8 @@ export function getInvoiceTypeCodeDescription(code: InvoiceTypeCode): string {
  * @param code The code to check
  */
 export function isValidInvoiceTypeCode(code: number): code is InvoiceTypeCode {
-    return code in InvoiceTypeCodes;
+    // return code in InvoiceTypeCodes;
+    return code > 1000
 }
 
 // Zod schema for invoice type codes (numeric codes)
